test(agents): isolate fd process failure fixture
